Adds localstorage of user to persist state across requests
The access token was stored in cookies but app needs to
access on frontend and allows access to token after page refresh
This commit sets token & user details in local storage on login. This can then
be retrieved and set in state for the auth context. As user &
token is set in context, child components can now access. Refresh endpoint is
queried to sign another access token when /verify is called.

Commit also adds more types for admin feature